+++ MauiProgram.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging;
+#if WINDOWS
+using Microsoft.Maui.LifecycleEvents;
+#endif
 
 namespace BattleshipMaui;
 
@@ -15,6 +18,19 @@ public static class MauiProgram
                 fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
                 fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
             });
+
+#if WINDOWS
+        builder.ConfigureLifecycleEvents(events =>
+        {
+            events.AddWindows(windows =>
+            {
+                windows.OnWindowCreated(window =>
+                {
+                    BattleshipMaui.WinUI.FullScreenHotkeyController.Attach(window);
+                });
+            });
+        });
+#endif
 
 #if DEBUG
         builder.Logging.AddDebug();